+++ src/Purchase/PurchaseContainer.js
@@ -63,13 +63,17 @@ class OrderPurchase extends React.Component {
     };
 
     deletePurchase = () => {
-        let {purchaseId} = this.state;
-        this.setState({isLoading: true});
-        this.props.deletePurchase(purchaseId)
-        setTimeout(()=>{
-            this.setState({isLoading: false})
-            this.handleModal();
-        }, 2000)
+        let {loggedInUser} = this.props;
+        const isAdmin = loggedInUser && loggedInUser.role_id === "admin";
+        if(isAdmin) {
+            let {purchaseId} = this.state;
+            this.setState({isLoading: true});
+            this.props.deletePurchase(purchaseId)
+            setTimeout(() => {
+                this.setState({isLoading: false})
+                this.handleModal();
+            }, 2000)
+        }
     }
 
     renderFooter = () => {
@@ -271,7 +275,8 @@ const mapStateToProps = (state, ownProps) => {
         product: state?.product?.products,
         purchase: state?.purchase?.purchase,
         loading: state?.client?.loading,
-        error: state?.client?.error
+        error: state?.client?.error,
+        loggedInUser: state?.loggedInUser?.data?.data,
     };
 };
 const mapDispatchToProps = (dispatch) => {
